test-run: avoid race conditions when issuing ports for new server instances
This is a stab at tarantool/gh-1163, ""Broken tarantool console handshake"
from test-run".

This fix attempts to remove a race condition between concurrent greenlets
of test-run, tarantoolctl and starting/stopping tarnatool server daemons
when issuing a port number.

This is a blind attempt to fix the problem without identifying the root
cause.