Parallelize/randomize per suite
Don't parallelize and don't randomize tests by default.

Add the following line to a `suite.ini` to mark suite ready to run in
parallel (as btw tests of the suite as well as at the time with other
suites) and with random test order:

```
is_parallel = True
```

Test-run will run all parallel-ready suites first, then all the rest.

There is a little 'race condition': suite marked as done when a first
worker done it's work (in other words when no more non-started test and
one of the tests done). The next worker run when some other done, so
it's possible that a first non-parallel suite will have a little
intersection by the running time periods with last tests of parallel
suite(s).