Allow to get rid of *.result files in TAP suites
When `*.result` file is not found for the test, its output checked to
being TAP-compatible ("ok ", "not ok " and so on) and in that case will
be checked as TAP output. The new result file won't created.

Checked app-tap, box-tap and sql-tap suites on c91a7145; the result
files are needed for the following tests:

* app-tap/init_script.test.lua    -- not TAP output
* app-tap/inspector.test.lua      -- not TAP output
* app-tap/pcall.test.lua          -- not TAP output
* box-tap/trigger_atexit.test.lua -- not TAP output
* box-tap/trigger_yield.test.lua  -- not TAP output
* app-tap/tap.test.lua            -- contains "not ok " which are okay
                                     for us.

Other `*-tap/*.result` files can be deleted.

The commit adds pytap13 dependency.

Fixes #57.