Support configurations in core = 'app' tests
The code was copy-pasted from tarantool_server.py to app_server.py and
adopted a bit, but it would be better to generalize it and move to
test_suite.py. I understand this point, but don't want to be beaten by
corner cases during proper refactoring (at least within this commit).