Don't abort if 'start server' command fails
Sometimes, it is useful to check that a tarantool replica fails to start
under certain conditions, but currently it is impossible to do so in
Lua, because test_run:cmd('start server') aborts the test in this case.
This patch makes it print an error message to the result file instead.

Closes #92