Allow to pass arguments to create_cluster.
Just like in case of "start server replica with args='1 2 3'", it is
now possible to pass arguments to create_cluster, so that all instances
are started with these arguments.

Also alter create_cluster a little bit: previous behaviour was to create
and start instances in a loop, change it to first creating all
instances and then starting them. This is needed because creating an
instance may take up a lot of time, if the old vardir isn't empty, and
the first started instance would have to wait for others for a
significant period of time, which makes it exit upon timeout if
replication_connect_timeout is relatively small.