Add new config param 'show_reproduce_content'
By default this param is True, if the test fails then the contents of
reproduce file are displayed. If the parameter is False, do not show
the contents of reproduce file.  For example, the 'sql-tap' tests run
each case in a separate tarantool instance, so reproduce is not
necessary for problem investigation. Used in 'suite.ini'.

Closes #113