Fix reporting of non-default server fail at start
Case on tarantool-2.2.0-107-g4c15e8860:

./test-run.py -j 1 box/transaction.test.lua box/admin.test.lua

The latter test do `test_run:cmd('restart server default')`, which
cannot start a server after the former test (see [1] for more info).

We need to check TarantoolStartError in LuaTest.execute(), because the
exception propagated to this greenlet if it occurs in an inspector
greenlet. I guess it is because this greenlet starts inspector and it
does not matter that it doesn't join on the inspector greenlet at the
moment.

[1]: https://github.com/tarantool/tarantool/issues/4093