Update pyyaml version
GitHub warns about pyyaml < 4.2b1, because of CVE-2017-18342:
yaml.load() could cause arbitrary code execution. I think it is not very
important in the case of test-run, because a test code is controlled by
a user, but anyway it is good to be up to date.

Eliminated new pyyaml warnings about using of unsafe yaml.load():
replaced it with yaml.safe_load(). We don't construct Python classes
directly according to yaml tags, so safe_load() fit our needs.