Catch non-default server start fail in app server
When test_run:cmd('start server foo') is called from an app test and
tarantool fails to start, the following exception is shown:

 | [010] Worker "010_app-tap" received the following error; stopping...
 | [010] Traceback (most recent call last):
 | [010]   File "/home/alex/projects/tarantool-meta/tarantool/test-run/lib/worker.py", line 294, in run_task
 | [010]     task, self.server, self.inspector)
 | [010]   File "/home/alex/projects/tarantool-meta/tarantool/test-run/lib/test_suite.py", line 208, in run_test
 | [010]     short_status = test.run(server)
 | [010]   File "/home/alex/projects/tarantool-meta/tarantool/test-run/lib/test.py", line 180, in run
 | [010]     self.execute(server)
 | [010]   File "/home/alex/projects/tarantool-meta/tarantool/test-run/lib/app_server.py", line 43, in execute
 | [010]     tarantool.join()

The commit handles this case and the exception should not be shown
anymore.

This commit follows the similar one for tarantool server: 7176cede ('Fix
reporting of non-default server fail at start ').

The reporting of such situations is not ideal for now (we show output
and logs for a default server, but don't do that for non-default one),
but this will be fixed in the scope of #159.

Fixes #115.