Fail testing when test-run fails internally
This commit handles the case when test-run found that some of scheduled
tasks were not reported as finished by workers. The old behaviour leads
to zero exit code in the case when --force option is not set (looks as
the unintended mistake).

Now this situation causes a testing fail (non-zero exit code) and it
does not matter whether --force option is set.

Show list of failed tasks always except the case when it is caused by a
keyboard interruption.

While we are here clarified and changed related test-run messages.

All that is about the situation that should never occur. See #115 for
such cases.