Add the signal option into 'stop server' command
How to use:

 | test_run:cmd('stop server foo with signal=KILL')

The 'stop server foo' command without the option sends SIGTERM as
before.

The reason of the commit is to provide the ability to stop a tarantool
instance even when it cannot be stopped with SIGTERM.

The example of such situation is when we're set ERRINJ_WAL_DELAY.
SIGTERM is not sufficient here, because the main thread waits for the
stuck WAL thread after the signal.