Fix false positive internal error detection
Tests are dispatched to task queue dispatchers at start and so it is
normal that tests can be not consumed when a test fails in non-force
mode or when a test hung. It is not test-run error.

The problem part with a failed (not hung) test was introduced in
d8abc031a6de3603b94660d66872f0485b5910cc ('Fail testing when test-run
fails internally').

Reported in
https://github.com/tarantool/test-run/issues/159#issuecomment-488065589