pretest_clean: clean strict module state
app/strict.test.lua expects that 'a' variable is not declared at the
start of the test. However if the strict mode is enabled (which is
default in a debug build) global variables assignments are tracked
inside a metatable of _G. A variable is considered declared even if it
is set to `nil`.

If a test that runs before app/strict.test.lua set 'a' to any value,
then app/strict.test.lua will fail.

So we need to clean the metatable of _G in addition to _G itself.