Send SIGKILL if server does not stop after timeout
If a tarantool instance does not handle SIGTERM correctly, it is the bug
in tarantool. A testing system should reveal problems rather than hide
them. However this patch doing that and I'll explain why.

Many tests we have perform instance starting / stopping: they are
written to spot different problems, but not this one. If everything else
except hanging at stop is okay, there is no reason to fail the test. We
print a notice on the terminal in the case and proceed further.

In future, when all problems of this kind will be resolved, we can
interpret presence of such notices as error. However now it'll not add
anything for quality of our testing.

We had a problem of this kind ([1]) in the past and now have another one
([2]). Until it will be resolved, it worth to handle the situation on
the testing system side.

[1]: https://github.com/tarantool/tarantool/issues/4127
[2]: https://github.com/tarantool/tarantool/issues/5573

Part of #157

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>