Postpone fragile tests
Added "fragile" option to suite.ini file format. The option lists tests
that are not desined be run in parallel with others: say, those ones
that lean of timings that are meet only when enough system resources are
available.

The option set in the same way as "disabled" and other test lists:

 | fragile = foo.test.lua ; gh-1234
 |           bar.test.lua ; gh-5678

Fixes #187.