Add --exclude option
The option allows to avoid execution of a certain subset of tests.

The option sets a pattern for exclusion and all tests are filtered using
it: if a full test name (say, app-tap/string.test.lua) contains the
pattern as a substring, then the test will be excluded from execution.

The option can be passed several times to set several exclusions
patterns.

A list of such exclusion patterns also can be provided using
TEST_RUN_EXCLUDE environment variable (patterns are separated by
whitespaces).

If both the environment variable and the option are provided, then the
environment variable value will be ignored.

Trick: `--exclude foo/` allows to exclude a test suite.

Fixes #54.