Handle incorrect test-run commands gracefully
I looked over places, where LuaPreprocessorException is raised and they
do not look as show-stoppers. Those situations are about incorrect
commands (say, parsing error) or commands evaluated in an incorrect
state (e.g. attempt to start a server that is not created yet). We can
return an error to the test, report a notice to the terminal and
continue the test. The decision, whether such cases should fail the
test, is up to a test author now.

The new format of displaying the notice on the terminal shifts
understanding of the error from 'looks as a testing system problem' to
'the test performs unexpected commands'.

Before the change such situations were reported to the terminal this
way:

 | [001] TarantoolInpector.handle() received the following error:
 | [001] <..a python stack trace and an error message..>
 | [001] LuaPreprocessorException: lua preprocessor error: "Can't stop nonexistent server 'foo'"

And the test stops.

After the change, it looks so:

 | [001] * [QA Notice]
 | [001] *
 | [001] * Can't stop nonexistent server 'foo'
 | [001] *

And the test receives a Lua error. Usually it leads to a fail of the
test, however now it is possible to catch the error with pcall() and
handle it.

The idea appears while I looking into [1]. The test creates a full-mesh
replicaset, spins in a cycle with stopping and starting instances and
then drops the replicaset. There are several situations that correctly
leads to failure of this particular test, but may be considered as
appropriate ones in another tests.

1. The replicaset instances are numbered from 1 to 3, but when the test
   is unable to determine an instance to stop during 60 seconds (default
   test_run:wait_cond() timeout), it tries to stop 0th instance. It is
   attempt to stop a nonexistent instance.
2. The similar to the problem 1, but with attempt to evaluate a command
   on a nonexistent instance.
3. An error may occurs between stopping and starting the instance. In
   this case the loop exits and test_run:drop_cluster() is unable to
   stop all instances in the replicaset.

The first problem is reported now as it should (with the QA Notice, as
shown above). The second will be reported this way after the next
commit. The third requires to ignore 'stop server <...>' failures in
test_run:drop_cluster() and will be handled in the separate commit
within this patchset.

While I'm here, changed several related things:

* Color 'info' messages as yellow. It affects several rarely shown
  existing messages, which looks okay.
* Removed the 'lua preprocessor error' prefix from the error message: it
  does not give any new information and so looks useless.
* Fixed conversion of the error to string: repr() -> str().
* Show an unhandled Lua error from a test-run command evaluation without
  a path to the test_run.lua file: it is useless and looks ugly (in
  particular a colon between the path and the error message obligates
  the yaml emitter to wrap the string into quotes and so duplicate all
  quotes inside).

[1]: https://github.com/tarantool/tarantool/issues/5395