Don't fail drop_cluster() on a stopped instance
Since the method is used as clean up after a test / a test case, it does
not look worthful to fail the test when one of instances is already
stopped. However test-run's QA notice will be printed to the terminal in
the case: it is possible that this situation is not expected and some
subtle problem exists. It worth to let a developer know about this.

Found while looking into [1].

[1]: https://github.com/tarantool/tarantool/issues/5395