logging: trace tarantool commands and responses
The commands and responses are written into test-run's log file that is
named like test/var/log/001_replication.log (where '001' is number of
worker and 'replication' is the test suite name).

A command is logged exactly as it is sent to tarantool: without newlines
in a middle[^1]. It'll look a bit weird for multiline commands, but may
ease debugging if a problem will be in test-run's way to format a
command before sending or somewhere around processing commands by the
console module.

[^1]: The last newline is stripped from logs for readability.

The idea appears while looking into [1].

[1]: https://github.com/tarantool/tarantool/issues/5395