logging: mark chunks without a newline at the end
test-run outputs some lines in several chunks, e.g.
"engine/tuple.test.lua", then "memtx", then "[ pass ]". When test-run
prints on the terminal, it accumulates the chunks and write them at once
to don't mix output from different workers. Logs cannot be bufferred
this way: otherwise we can miss information that is valuable for
debugging.

After the previous commit ('logging: prepend each log entry with a
timestamp') we write each of such chunks on its own line and they looks
as if each would be a full line. In order to distinguish a chunk from a
full line, add `<no newline>` after a chunk without a newline at the
end.

The difference does not look important, but it should make easier to
understand relation between strings shown on the terminal and ones
written to the test-run log file.

Part of activities around [1].

[1]: https://github.com/tarantool/tarantool/issues/5395