Don't reset no-output timer on irrelevant events
The idea of this timer is to stop testing if nothing is written to the
terminal for a long time (120 seconds by default).

Before this commit, the timer reset on any event: say, when an entry is
written to a test-run log file. It is counter-intuitive and sometimes
may look like the timer just does not work.

Let's consider the following snippet of a test code:

 | -- Wait until <cmd> on instance 'foo' will return
 | -- an expected value.
 | local cmd = <...>
 | local exp = <...>
 | while true do                                \
 |     local res = test_run:eval('foo', cmd)[1] \
 |     if res == exp then                       \
 |         break                                \
 |     end                                      \
 | end

After commit 'logging: trace test-run commands and responses' we log
test-run commands and so the snippet would discard the no-output timer
at all. If something going wrong and the <cmd> command will never return
the expected result, the loop will spin infinitely.

After this commit, the logging events will not reset the timer and so
the testing will be finished eventually (after 120 seconds without
output to the terminal).

However if --debug is passed, the logging events will be printed on the
terminal and will reset the timer. It looks logical, because the timer
is about the printed messages, but it means that the behaviour of the
testing system may depend on whether the --debug option is passed. Don't
know whether behaviour would be better, so implemented the most
intuitive one.

Part of activities around [1].

[1]: https://github.com/tarantool/tarantool/issues/5395