Explicit imports of needed objects
There are two import approaches in Python: import the whole module and
import only needed objects (class, variable etc). We don't have code a
style guide for test-run project and I personally like second approach.
Moreover it seems that 'from foo import bar' is the preferred way within
the test-run's code.

This patch has been made by Alexander Turenko in scope of task to make
code compatible with Python 3, but actually it is not related to it.

Part of #20

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>