Explicit imports of needed objects
There are two import approaches in Python: import everything from a module and
import only needed objects (class, variable etc). According to [1] and [2]
second approach allows to keep code clean for readers. This patch has been made
by Alexander Turenko in scope of task to make code compatible with Python 3,
but actually it is not related to it.

1. https://s16h.medium.com/importing-star-in-python-88fe9e8bd4d2
2. https://softwareengineering.stackexchange.com/questions/187403/import-module-vs-from-module-import-function

Part of #20

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>