Increase no-output-timeout default value
Found that issue with hanging tests on 'restart server deafult' and
common commands, was because of the not enough 'no-output-timeout'
timeout, which caused tests to be broken. It happened because to many
tests were run in parallel and it caused some tests to be stopped with
the timeout. Decided to set 20 minutes timeout instead of previous 2
minutes.