Add --bootstrap option for accept a bootstrap snap
The idea is the same as for --snapshot, but src/box/bootstrap.snap can
be passed here. Added checks whether a snapshot aims to be used for
bootstrap or for local recovery.

A given bootstrap snapshot is prepared to use with local recovery. We
have no other way to feed an external bootstrap snapshot to tarantool
until [1] will be implemented.

[1]: https://github.com/tarantool/tarantool/issues/5540