refactoring: provide internal qa_notice() function
Now we have only one place, where a [QA Notice] message is printed. But
we'll have more and it is convenient to have one function for this
purpose.

The function should be used in places, where we observe an incorrect
behaviour of tarantool or a test, but want to lower it to a warning and
proceed further. Of course, there should be a solid reason to do so.

Part of #157