Set PWD env variable to a worker working directory
test-run copies .tarantooctl file from a test/ directory to a worker's
working directory (e.g. test/var/001_app). It would be quite intuitive
if tarantoolctl would use this copy when operating on tarantool
instances of the worker. Suprisingly, it is not so.

tarantoolctl checks whether the file exists inside a directory pointed
by the PWD environment variable, not a real current working directory.
test-run did not set PWD prior to this commit, however everything work.
Why? Because test-run.py is always executed from the test/ directory,
which contains the same .tarantoolctl file. Even `make test` / `make
test-force` targets issue commands like `cd <...>/test &&
<...>/test-run.py <...>` (where `cd` sets PWD).

The time has come. Let's set PWD and use test/var/001_app/.tarantoolctl
configuration for instances of the 001_app worker.

Side note: it would be good to fix tarantoolctl too to use a real cwd,
not PWD.

This inconsistency was spotted, when we perform attempt to remove the
test/.tarantoolctl config from the main tarantool repository, move it to
the test-run repository, but keep copying into a worker directory.

As a side effect, test-run now could be invoked without manual changing
a current directory to test/. So, where we run a command like this
(example for in-source build):

 | (cd test && ./test-run.py)

The following command can be used:

 | ./test/test-run.py

Nice!

Sadly, the trick `Popen(<...>, env=dict(os.environ, PWD=self.vardir))`
does not work for us, because we use os.putenv() in test-run and Python
tests. os.putenv() does not update os.environ, so we would feed
non-actual environment to the subprocess. Because of this, we just set
PWD before and after Popen() call.

Part of tarantool/tarantool#5504
Part of #78

Co-authored-by: Alexander V. Tikhonov <avtikhon@tarantool.org>