Wait until residual servers will be stopped
Otherwise, if a next test will try to start a new instance with the
same name, it may fail with the 'The daemon is already running' error,
if a previous test fails or reaches --test-timeout.

The commit fixes the problem for 'core = tarantool' test suites. The
next commit will resolve it for 'core = app' test suites.

See a test case in PR #244 (app/test-timeout.test.lua).

A few words about details of the implementation. The killall_servers()
method performs almost the same actions that stop_nondefault(), so I
kept the latter and removed the former. However there are differences:

- killall_servers() does not wait for termination of processes. This is
  why the change fixes given problem.
- killall_servers() sends SIGKILL, while stop_nondefault() stop
  instances as usual: SIGTERM, wait 5 seconds at max, SIGKILL, wait for
  termination. I'll return instant SIGKILL in one of the following
  commits: I splitted the changes for ease reading.

Part of #157