Stop residual instances after an app test
If a test fails or reaches --test-timeout, some non-default tarantool
instances may be left running and so if a next test starts an instance
with the same name, it may fail with the 'The daemon is already running'
error.

This change follows the approach of the previous commit ('Wait until
residual servers will be stopped').

See a test case in PR #244 (app-tap/test-timeout.test.lua).

Part of #65
Part of #157