Handle 'No such process' exception
`send_signal()`, `terminate()` or `kill()` methods of a subprocess.Popen
object raise OSError on Python 2, when the process does not exist and
Python knows about it. See:

 | >>> import subprocess
 | >>> p = subprocess.Popen(['true'])
 | >>> p.poll()
 | 0
 | >>> p.kill()
 | Traceback (most recent call last):
 |   File "<stdin>", line 1, in <module>
 |   File "/usr/lib64/python2.7/subprocess.py", line 1279, in kill
 |     self.send_signal(signal.SIGKILL)
 |   File "/usr/lib64/python2.7/subprocess.py", line 1269, in send_signal
 |     os.kill(self.pid, sig)
 | OSError: [Errno 3] No such process

Python 3 does not raise the exception in the case. Let's protect ourself
from possible races / behaviour differences and silence the exception.

Part of #157