Relax dependency version requirements
It simplifies dependency managements.

Example: PyYAML 5.4.1 (most recent at the moment of writting this) is
available for Python 3.6+. However we should keep 5.3.1 for Python 3.5.
Anyway, it would be nice to depend on a newer (but compatible) version
on newer python versions to bring bugfixes and security updates. See
more in [1].

[1]: https://github.com/tarantool/test-run/pull/282

Supersedes PR #282.